LPD-92680 Move CryptoException to its own exception package
Mirror the SecretException change: give CryptoException a dedicated exception package and export it, rather than keeping it beside the manager API.
