+++ modules/apps/portal-security/portal-security-key-api/bnd.bnd
@@ -3,4 +3,5 @@ Bundle-SymbolicName: com.liferay.portal.security.key.api
 Bundle-Version: 1.0.2
 Export-Package:\
 	com.liferay.portal.security.key,\
-	com.liferay.portal.security.key.crypto
+	com.liferay.portal.security.key.crypto,\
+	com.liferay.portal.security.key.crypto.exception
