+++ modules/apps/portal-security/portal-security-key-api/src/main/java/com/liferay/portal/security/key/crypto/exception/CryptoException.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.security.key.crypto;
+package com.liferay.portal.security.key.crypto.exception;
 
 import com.liferay.portal.kernel.exception.PortalException;
