+++ modules/apps/portal-security/portal-security-key-api/src/main/resources/com/liferay/portal/security/key/crypto/exception/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
