+++ modules/apps/portal-security/portal-security-key-impl/src/test/java/com/liferay/portal/security/key/internal/crypto/CryptoManagerImplTest.java
@@ -12,8 +12,8 @@ import com.liferay.portal.kernel.test.ReflectionTestUtil;
 import com.liferay.portal.kernel.test.util.RandomTestUtil;
 import com.liferay.portal.security.key.KeyReference;
 import com.liferay.portal.security.key.ServiceIndicator;
-import com.liferay.portal.security.key.crypto.CryptoException;
 import com.liferay.portal.security.key.crypto.CryptoServiceResult;
+import com.liferay.portal.security.key.crypto.exception.CryptoException;
 import com.liferay.portal.security.key.spi.ProviderStatus;
 import com.liferay.portal.security.key.spi.crypto.CryptoProvider;
 import com.liferay.portal.security.key.spi.profile.KeyManagerProfile;
